Ajustado cadastro e remoção de atividades

+++ src/app/services/service_atividades.ts
@@ -4,7 +4,7 @@ import { from, map, Observable } from 'rxjs';
 import { Atividade } from '../interfaces/atividades';
 import { PDFDocument, rgb, StandardFonts } from 'pdf-lib';
 import { SupabaseService } from '../core/services/SupaBaseService';
-import { adaptarAtividadeParaRequest, adaptarAtividadeParaResponse } from '../shared/adapters/ativdade.adapter';
+import { adaptarAtividadeParaRequest, adaptarAtividadeParaResponse } from '../shared/adapters/atividade.adapter';
 import { adaptarTurmaParaRequest } from '../shared/adapters/turma.adapter';
 
 @Injectable({
@@ -22,7 +22,10 @@ export class ServiceAtividades {
     return from(
       this.supabase.getClient()
         .from(this.tabela)
-        .select('*')
+        .select(
+          `*,
+          materia (nome)`
+        )
     ).pipe(
       map(({ data, error }) => {
         if (error) throw error;
@@ -34,19 +37,42 @@ export class ServiceAtividades {
   cadastrarAtividade(atividade: Atividade): Observable<Atividade> {
     atividade.user_id = this.supabase.getUserId();
 
-    const payload = adaptarAtividadeParaRequest(atividade);
+    const bucket = 'atividades'
     return from(
-      this.supabase.getClient()
-        .from(this.tabela)
-        .insert(payload)
-        .select()
-        .single()
-    ).pipe(
-      map(({ data, error }) => {
+      (async () => {
+        if (atividade.arquivo) {
+          const extensao = atividade.arquivo.type.split('/')[1];
+          const fileName = `${atividade.codigo}-${crypto.randomUUID()}.${extensao}`;
+          const filePath = `${fileName}`;
+
+          const { error: uploadError } = await this.supabase.getClient()
+            .storage
+            .from(bucket)
+            .upload(filePath, atividade.arquivo, {
+              cacheControl: '1000',
+              upsert: false
+            });
+
+          if (uploadError) throw uploadError;
+
+          atividade.url = filePath;
+          atividade.extensao = extensao;
+          atividade.nomeArquivo = fileName;
+
+        }
+
+        const payload = adaptarAtividadeParaRequest(atividade);
+
+        const { data, error } = await this.supabase.getClient()
+          .from(this.tabela)
+          .insert(payload)
+          .select()
+          .single();
+
         if (error) throw error;
         return adaptarAtividadeParaResponse(data);
-      })
-    );
+      })()
+    )
   }
 
   async adicionarTextoRodapePDF(pdfBytes: Uint8Array): Promise<Uint8Array> {
@@ -88,9 +114,13 @@ export class ServiceAtividades {
 
   findById(id: Number): Observable<Atividade> {
     return from(
+
       this.supabase.getClient()
         .from(this.tabela)
-        .select('*')
+        .select(
+          `*,
+          materia (nome)`
+        )
         .eq('id', id)
         .single()
     ).pipe(
@@ -101,23 +131,38 @@ export class ServiceAtividades {
     );
   }
 
-  //  findArquivoByAtividade(id: number): Observable<Blob> {
-  //    return this.http.get(`${this.URL}/${id}/arquivo`, { responseType: 'blob' })
-  //  }
 
   removerAtividadeById(id: number): Observable<Atividade> {
     return from(
-      this.supabase.getClient()
-        .from(this.tabela)
-        .delete()
-        .eq('id', id)
-        .select()
-        .single()
-    ).pipe(
-      map(({ data, error }) => {
-        if (error) throw error;
-        return adaptarAtividadeParaResponse(data);
-      })
+      (async () => {
+        const { data: atividade, error: getError } = await this.supabase.getClient()
+          .from(this.tabela)
+          .select('*')
+          .eq('id', id)
+          .single()
+
+        if (getError) throw getError;
+
+        if (atividade?.arquivo_anexado && atividade?.url) {
+          const { error: storageError } = await this.supabase.getClient()
+            .storage
+            .from('atividades')
+            .remove([atividade.url]);
+
+          if (storageError) throw storageError;
+        }
+
+        const { data, error } = await this.supabase.getClient()
+          .from(this.tabela)
+          .delete()
+          .eq('id', id)
+          .select()
+          .single();
+
+        if (error) throw error
+
+        return adaptarAtividadeParaResponse(data)
+      })()
     );
   }
 }
